otp error

+++ app/src/main/java/com/practies/groceryapp/repository/GroceryRepository.kt
@@ -1,5 +1,7 @@
 package com.practies.groceryapp.repository
 
+import com.practies.groceryapp.model.SignInData
+import com.practies.groceryapp.model.SignUpData
 import com.practies.groceryapp.network.ApiService
 import javax.inject.Inject
 
@@ -8,6 +10,15 @@ class GroceryRepository @Inject constructor(private val apiService: ApiService) 
 
     suspend fun getAllProducts()=apiService.getAllProducts()
 
-   suspend fun  generateOtp(phone:Int)=apiService.generateOtp(phone)
+   // suspend fun getProductDetails(productId:Int)=apiService.getProductDetails(productId)
+
+    suspend fun  generateOtp(phone:Long)=apiService.generateOtp(phone)
+
+
+   // suspend fun userSignUp(signUpData: SignUpData)=apiService.signUp(signUpData)
+
+  //  suspend fun  userSignIn(signInData: SignInData)=apiService.signIn(signInData)
+
+   // suspend fun  checkPhoneNumber
 
 }
